requirements.txt
python-chess
kivy